Refactoring, code cleanups, bugfixes, use modern Java features

+++ src/main/java/com/cburch/draw/shapes/Line.java
@@ -234,7 +234,7 @@ public class Line extends AbstractCanvasObject {
     if (attr == DrawAttr.STROKE_COLOR) {
       strokeColor = (Color) value;
     } else if (attr == DrawAttr.STROKE_WIDTH) {
-      strokeWidth = ((Integer) value).intValue();
+      strokeWidth = (Integer) value;
     }
   }
 }
