+++ src/main/java/com/cburch/logisim/gui/menu/ProjectCircuitActions.java
@@ -331,7 +331,7 @@ public class ProjectCircuitActions {
     dlog.setVisible(true);
     field.requestFocusInWindow();
     Object action = pane.getValue();
-    if (!(action instanceof Integer) || ((Integer) action).intValue() != OptionPane.OK_OPTION) {
+    if (!(action instanceof Integer) || (Integer) action != OptionPane.OK_OPTION) {
       return null;
     }
