+++ src/main/java/com/cburch/logisim/gui/opts/MouseOptions.java
@@ -268,7 +268,7 @@ class MouseOptions extends OptionsPanel {
     public Object getValueAt(int row, int column) {
       Integer key = cur_keys.get(row);
       if (column == 0) {
-        return InputEventUtil.toDisplayString(key.intValue());
+        return InputEventUtil.toDisplayString(key);
       } else {
         Tool tool = getOptions().getMouseMappings().getToolFor(key);
         return tool.getDisplayName();
