+++ src/main/java/com/cburch/logisim/std/arith/BitAdder.java
@@ -81,7 +81,7 @@ public class BitAdder extends InstanceFactory {
 
   private void configurePorts(Instance instance) {
     BitWidth inWidth = instance.getAttributeValue(StdAttr.WIDTH);
-    int inputs = instance.getAttributeValue(NUM_INPUTS).intValue();
+    int inputs = instance.getAttributeValue(NUM_INPUTS);
     int outWidth = computeOutputBits(inWidth.getWidth(), inputs);
 
     int y;
@@ -113,7 +113,7 @@ public class BitAdder extends InstanceFactory {
 
   @Override
   public Bounds getOffsetBounds(AttributeSet attrs) {
-    int inputs = attrs.getValue(NUM_INPUTS).intValue();
+    int inputs = attrs.getValue(NUM_INPUTS);
     int h = Math.max(40, 10 * inputs);
     int y = inputs < 4 ? 20 : (((inputs - 1) / 2) * 10 + 5);
     return Bounds.create(-40, -y, 40, h);
@@ -148,7 +148,7 @@ public class BitAdder extends InstanceFactory {
   @Override
   public void propagate(InstanceState state) {
     int width = state.getAttributeValue(StdAttr.WIDTH).getWidth();
-    int inputs = state.getAttributeValue(NUM_INPUTS).intValue();
+    int inputs = state.getAttributeValue(NUM_INPUTS);
 
     // compute the number of 1 bits
     int minCount = 0; // number that are definitely 1
